refactor: NoticeList 이름 변경 및 불필요한 변수 제거
